# projects/dunkware-street-model/api/openapi.yaml
openapi: 3.0.6
info:
  description: Schema for Dunkware Street Model components.
  title: Dunkware Street Model
  version: 1.1.0
servers:
- description: Production server
  url: https://api.dunkware.com/v1
paths:
...
  /api/exchanges/{exchangeId}/tickers:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TickerRef'
                type: array
          description: A list of tickers for the exchange
        "404":
          description: Exchange not found
        "500":
          description: Internal server error
      summary: Get all tickers for a specific exchange
      x-accepts:
      - application/json
  /api/exchanges/{exchangeId}/tickers/{tickerId}/variables:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the ticker
        explode: false
        in: path
        name: tickerId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TickerVariable'
                type: array
          description: A list of variables for the ticker
        "404":
          description: Ticker or Exchange not found
        "500":
          description: Internal server error
      summary: Get all variables for a specific ticker
      x-accepts:
      - application/json
  /api/exchanges/{exchangeId}/sessions:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ExchangeSession'
                type: array
          description: A list of sessions for the exchange
        "404":
          description: Exchange not found
        "500":
          description: Internal server error
      summary: Get all sessions for a specific exchange
      x-accepts:
      - application/json
  /api/exchanges/{exchangeId}/tickers/{tickerId}/sessions:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the ticker
        explode: false
        in: path
        name: tickerId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TickerSession'
                type: array
          description: A list of sessions for the ticker
        "404":
          description: Ticker or Exchange not found
        "500":
          description: Internal server error
      summary: Get all sessions for a specific ticker
      x-accepts:
      - application/json
  /api/exchanges/{exchangeId}/tickers/{tickerId}/sessions/{sessionId}/stats:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the ticker
        explode: false
        in: path
        name: tickerId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the ticker session
        explode: false
        in: path
        name: sessionId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TickerStat'
                type: array
          description: A list of stats for the ticker session
        "404":
          description: "Ticker, Exchange, or Session not found"
        "500":
          description: Internal server error
      summary: Get all stats for a specific ticker session
      x-accepts:
      - application/json
  /api/exchanges/{exchangeId}/tickers/{tickerId}/stats/{statId}:
    get:
      parameters:
      - description: Numeric identifier of the exchange
        explode: false
        in: path
        name: exchangeId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the ticker
        explode: false
        in: path
        name: tickerId
        required: true
        schema:
          type: integer
        style: simple
      - description: Numeric identifier of the stat
        explode: false
        in: path
        name: statId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TickerStat'
          description: Details of the specific stat for the exchange ticker
        "404":
          description: "Ticker, Exchange, or Stat not found"
        "500":
          description: Internal server error
      summary: Get a specific stat for a specific exchange ticker
      x-accepts:
      - application/json
components:
  schemas:
...
    TickerRef:
      properties:
        id:
          description: Numeric identifier for the ticker
          type: integer
        identifier:
          description: String identifier for the ticker
          type: string
        exchange:
          description: Numeric identifier for the exchange
          type: integer
        name:
          description: Name of the ticker
          type: string
        sessionCount:
          description: Number of sessions associated with the ticker
          type: integer
        baseVersion:
          description: Base version of the ticker
          format: double
          type: number
      required:
      - baseVersion
      - exchange
      - id
      - identifier
      - name
      - sessionCount
      type: object
    TickerVariable:
      properties:
        id:
          description: Numeric identifier for the ticker variable
          type: integer
        identifier:
          description: String identifier for the ticker variable
          type: string
        name:
          description: Name of the ticker variable
          type: string
        group:
          description: Group to which the ticker variable belongs
          type: string
        dataType:
          $ref: '#/components/schemas/GenDataType'
        dataFormat:
          $ref: '#/components/schemas/GenDataFormat'
        baseVersion:
          description: Base version of the ticker variable
          format: double
          type: number
        sessionCount:
          description: Number of sessions associated with the ticker variable
          type: integer
        value:
          $ref: '#/components/schemas/TickerVariable_value'
      required:
      - baseVersion
      - dataFormat
      - dataType
      - group
      - id
      - identifier
      - name
      - sessionCount
      type: object
    ExchangeSession:
      properties:
        id:
          description: Numeric identifier for the exchange session
          type: integer
        exchangeId:
          description: Numeric identifier for the exchange
          type: integer
        startTime:
          $ref: '#/components/schemas/GenTime'
        stopTime:
          $ref: '#/components/schemas/GenTime'
        date:
          $ref: '#/components/schemas/GenDate'
        entityCount:
          description: Number of entities associated with the session
          type: integer
        version:
          description: Version of the exchange session
          format: double
          type: number
      required:
      - date
      - entityCount
      - exchangeId
      - id
      - startTime
      - stopTime
      - version
      type: object
    TickerSession:
      properties:
        date:
          $ref: '#/components/schemas/GenDate'
        startTime:
          $ref: '#/components/schemas/GenTime'
        stopTime:
          $ref: '#/components/schemas/GenTime'
        signalCount:
          description: Number of signals associated with the session
          type: integer
        version:
          description: Version of the ticker session
          format: double
          type: number
      required:
      - date
      - signalCount
      - startTime
      - stopTime
      - version
      type: object
    TickerStat:
      properties:
        statId:
          description: Numeric identifier for the statistic
          type: integer
        statName:
          description: Name of the statistic
          type: string
        tickerId:
          description: Numeric identifier for the ticker
          type: integer
        exchangeId:
          description: Numeric identifier for the exchange
          type: integer
        elementId:
          description: Numeric identifier for the element
          type: integer
        statValue:
          description: Value of the statistic
          format: double
          type: number
        time:
          $ref: '#/components/schemas/GenTime'
      required:
      - elementId
      - exchangeId
      - statId
      - statName
      - statValue
      - tickerId
      type: object
...
    GenDate:
      example:
        year: 2024
        month: 5
        dayOfMonth: 25
      properties:
        year:
          description: The year component of the date.
          type: integer
        month:
          description: The month component of the date.
          type: integer
        dayOfMonth:
          description: The day of the month component.
          type: integer
      required:
      - dayOfMonth
      - month
      - year
      type: object
    GenTime:
      example:
        hour: 14
        minute: 30
        second: 0
        timeZone: UTC
      properties:
        hour:
          description: The hour component of the time.
          type: integer
        minute:
          description: The minute component of the time.
          type: integer
        second:
          description: The second component of the time.
          type: integer
        timeZone:
          description: Optional string representing the time zone.
          type: string
      required:
      - hour
      - minute
      - second
      type: object
...
    GenDateTime:
      example:
        date:
          year: 2024
          month: 5
          dayOfMonth: 25
        time:
          hour: 14
          minute: 30
          second: 0
          timeZone: UTC
      properties:
        date:
          $ref: '#/components/schemas/GenDate'
        time:
          $ref: '#/components/schemas/GenTime'
      type: object
...
    GenDataType:
      description: Enumeration of generic data types.
      enum:
      - Integer
      - Double
      - String
      - Boolean
      - GenTime
      - GenDate
      - GenDateTime
      example: Integer
      type: string
    GenDataFormat:
      example:
        formatType: roundToDecimalPlaces
        precision: 2
      properties:
        formatType:
          description: The type of data formatting to apply.
          enum:
          - roundToDecimalPlaces
          - displayAsPercent
          - addCommas
          - truncateString
          - currency
          - abbreviateNumber
          type: string
        precision:
          description: Number of decimal places for rounding (used with roundToDecimalPlaces).
          type: integer
        truncateLength:
          description: Maximum length for truncating strings (used with truncateString).
          type: integer
        currencySymbol:
          description: Symbol for currency formatting (used with currency).
          type: string
        abbreviationStyle:
          description: Style for abbreviating large numbers (used with abbreviateNumber).
          type: string
      type: object
...
    TickerVariable_value:
      description: "Value of the ticker variable, type depends on the dataType field"
      oneOf:
      - type: integer
      - type: number
      - type: string
      - type: boolean
      - $ref: '#/components/schemas/GenTime'
      - $ref: '#/components/schemas/GenDate'
      - $ref: '#/components/schemas/GenDateTime'
...