# projects/dunkware-street-server/Users/duncankrebs/dunkhub/street-stack/projects/dunkware-street-server/src/main/java/com/dunkware/street/model/GenDataType.java
package com.dunkware.street.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonValue;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;


import java.util.*;
import jakarta.annotation.Generated;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

/**
 * Enumeration of generic data types.
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-06-03T04:43:50.832023-05:00[America/Chicago]")
public enum GenDataType {
  
  INTEGER("Integer"),
  
  DOUBLE("Double"),
  
  STRING("String"),
  
  BOOLEAN("Boolean"),
  
  GENTIME("GenTime"),
  
  GENDATE("GenDate"),
  
  GENDATETIME("GenDateTime");

  private String value;

  GenDataType(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static GenDataType fromValue(String value) {
    for (GenDataType b : GenDataType.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}